// backend/src/types/billings.ts
export enum TipoPlano {
  MENSAL = "MENSAL",
  POR_USO = "POR_USO",
  PERSONALIZADO = "PERSONALIZADO",
  POR_VAGA = "POR_VAGA",
  PERCENTUAL = "PERCENTUAL",
}

export enum CategoriaPlano {
  PLATAFORMA = "PLATAFORMA",
  RECRUTAMENTO_COM_RQE = "RECRUTAMENTO_COM_RQE",
  RECRUTAMENTO_SEM_RQE = "RECRUTAMENTO_SEM_RQE",
  RECRUTAMENTO_DIVERSOS = "RECRUTAMENTO_DIVERSOS",
}